Separate out Login into own component
Use a component for Login with proper styling and logo
